requirements.txt
aiogram==2.25.1
huggingface-hub
imageio
numpy
Pillow
torch
torchvision
git+https://github.com/sberbank-ai/Real-ESRGAN.git
